Create modular dalamud_switcher package
