+++ README.md
@@ -1,2 +1,19 @@
-# DalamudVersionSwitcher
- 
+# Dalamud Version Switcher
+
+A small Tkinter utility for toggling Dalamud beta mode for FFXIV.
+
+## Usage
+
+Run the application using the module entry point:
+
+```bash
+python -m dalamud_switcher
+```
+
+## Building
+
+Use the included batch script to create a standalone executable with PyInstaller:
+
+```bat
+build.bat
+```
